refactor: remove all temporary barrel files, migrate consumers to direct imports
Delete 3 barrel files that were marked for future removal:
- src/orchestrator/agent/tools_v2.py (6 consumers migrated)
- frontend/src/components/command-center/presentation.tsx (1 consumer)
- frontend/src/components/sidebar/panels.tsx (1 consumer)

All imports now reference canonical module locations directly.
Backend: 945 passed. Frontend: tsc clean, build successful.
